researcher/core/researcher.py
"""Main researcher class implementation."""

import os
import sys
# Add the agents package to the Python path
agents_path = os.path.abspath(os.path.join(os.path.dirname(__file__), "../../../agents/src"))
sys.path.append(agents_path)

from typing import List, Optional
from pydantic import BaseModel
from researchagent.agents import CodeAgent
from researchagent.models import LiteLLMModel

class ResearchResult(BaseModel):
    title: str
    content: str
    source: str
    relevance_score: float


class Researcher:
    """Main researcher class that handles research operations."""
    
    def __init__(self):
        self.results: List[ResearchResult] = []
        self.llm = LiteLLMModel(
            model_id="deepseek/deepseek-chat",
        )
        self.init_agent()

    def init_agent(self):
        self.agent = CodeAgent(
            model=self.llm,
            tools=[],
        )

    async def research(self, query: str) -> List[ResearchResult]:
        """
        Conduct research based on the given query.
        
        Args:
            query: The research question or topic
            
        Returns:
            List of research results
        """
        
        return self.results
    
    async def summarize(self, results: Optional[List[ResearchResult]] = None) -> str:
        """
        Summarize research results.
        
        Args:
            results: Optional list of results to summarize. If None, uses stored results
            
        Returns:
            Summary of the research
        """
        # TODO: Implement summarization logic
        return "Research summary will be implemented here."
