pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "researcher"
version = "0.1.0"
authors = [
    { name = "Eeza", email = "" },
]
description = "An AI-powered research assistant tool for automated information gathering and analysis"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "researchagents @ git+https://github.com/igu1/research-agents.git",
    "langchain>=0.0.200",
    "openai>=0.27.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "pydantic>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "aiohttp>=3.8.0",
    "tiktoken>=0.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[project.urls]
"Homepage" = "https://github.com/igu1/research-agents"
"Bug Tracker" = "https://github.com/igu1/research-agents/issues"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
asyncio_mode = "auto"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
